CHAR_MatchingLobby 삭제, ADPCharactor로 합병
