+++ proj_a/Source/proj_a/Component/Public/DPWeaponActorComponent.h
@@ -42,8 +42,6 @@ public:
 	UFUNCTION(BlueprintCallable, Category = "Weapon")
 	void AddWeapons(TSubclassOf<ADPWeapon> weaponClass);
 	UFUNCTION(BlueprintCallable, Category = "Weapon")
-	void AddWeapons_MatchLobby(TSubclassOf<ADPWeapon> weaponClass);
-	UFUNCTION(BlueprintCallable, Category = "Weapon")
 	void Equip(TSubclassOf<ADPWeapon> weaponClass);
 	UFUNCTION(BlueprintCallable, Category = "Weapon")
 	bool Attack(ADPPlayerController* controller, FHitResult& result, FRotator& info);
